+++ src/main/java/com/equipeturma862/cadastronf/repository/AgenciaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AgenciaRepository extends CrudRepository<Agencia, Integer > {
+public interface AgenciaRepository extends CrudRepository<Agencia, Long > {
 }
