+++ src/main/java/com/equipeturma862/cadastronf/repository/NotasFiscaisRepositoy.java
@@ -3,5 +3,5 @@ package com.equipeturma862.cadastronf.repository;
 import com.equipeturma862.cadastronf.domain.NotaFiscal;
 import org.springframework.data.repository.CrudRepository;
 
-public interface NotasFiscaisRepositoy extends CrudRepository<NotaFiscal, Integer> {
+public interface NotasFiscaisRepositoy extends CrudRepository<NotaFiscal, Long> {
 }
