+++ src/main/java/com/equipeturma862/cadastronf/repository/RemetenteRepositoy.java
@@ -3,5 +3,5 @@ package com.equipeturma862.cadastronf.repository;
 import com.equipeturma862.cadastronf.domain.Remetente;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RemetenteRepositoy extends CrudRepository<Remetente, Integer> {
+public interface RemetenteRepositoy extends CrudRepository<Remetente, Long> {
 }
