+++ src/main/java/com/equipeturma862/cadastronf/repository/RequisicaoRepository.java
@@ -3,5 +3,5 @@ package com.equipeturma862.cadastronf.repository;
 import com.equipeturma862.cadastronf.domain.Requisicao;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RequisicaoRepository extends CrudRepository<Requisicao, Integer> {
+public interface RequisicaoRepository extends CrudRepository<Requisicao, Long> {
 }
