+++ src/main/java/com/equipeturma862/cadastronf/service/AgenciaService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface AgenciaService {
     List<Agencia> list(String nome);
     Agencia save(Agencia agencia);
-    Agencia getById(Integer id);
-    Agencia update(Integer id, Agencia agencia);
-    void delete(Integer id);;
+    Agencia getById(Long id);
+    Agencia update(Long id, Agencia agencia);
+    void delete(Long id);;
 }
